fix : 핫 키워드 조회 버그 수정
