+++ shorts-domain/src/main/kotlin/com/mashup/shorts/domain/news/NewsRepository.kt
@@ -1,5 +1,6 @@
 package com.mashup.shorts.domain.news
 
+import java.time.LocalDateTime
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 import com.mashup.shorts.domain.category.Category
@@ -8,6 +9,11 @@ import com.mashup.shorts.domain.category.Category
 interface NewsRepository : JpaRepository<News, Long>, NewsQueryDSLRepository {
 
     fun findAllByCategory(category: Category): List<News>
-    fun findByTitleAndNewsLinkAndPress(title: String, newsLink: String, press: String): List<News>
     fun findTopByOrderByIdDesc(): News?
+
+    fun findAllByCreatedAtBetween(
+        startDateTime: LocalDateTime,
+        endDateTime: LocalDateTime,
+    ): List<News>
+
 }
