+++ shorts-domain/src/main/kotlin/com/mashup/shorts/domain/news/NewsRetrieve.kt
@@ -1,5 +1,7 @@
 package com.mashup.shorts.domain.news
 
+import java.time.LocalDateTime
+import java.time.LocalTime
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
@@ -11,8 +13,8 @@ import com.mashup.shorts.domain.membernews.MemberNewsRepository
 @Service
 @Transactional(readOnly = true)
 class NewsRetrieve(
-    private val newsRepository: NewsRepository,
     private val memberNewsRepository: MemberNewsRepository,
+    private val newsRepository: NewsRepository,
 ) {
 
     fun retrieveNews(member: Member, newsId: Long): NewsRetrieveInfo {
@@ -24,6 +26,33 @@ class NewsRetrieve(
             return NewsRetrieveInfo(newsLink = news.newsLink, isSaved = true)
         }
         return NewsRetrieveInfo(newsLink = news.newsLink, isSaved = false)
+    }
+
+    fun retrieveByHotKeyword(
+        targetDateTime: LocalDateTime,
+        keyword: String,
+        cursorId: Long,
+        size: Int,
+    ): List<News> {
+        val startDateTime = LocalDateTime.of(
+            targetDateTime.toLocalDate(),
+            LocalTime.of(targetDateTime.hour, 0)
+        )
+
+        val endDateTime = LocalDateTime.of(
+            targetDateTime.toLocalDate(),
+            LocalTime.of(targetDateTime.hour, 59)
+        )
+        val newsBundle = newsRepository.findAllByCreatedAtBetween(startDateTime, endDateTime)
+        val newsIds = newsBundle.map { it.id }
+
+        return newsRepository.loadNewsBundleByCursorIdAndTargetTime(
+            startDateTime,
+            endDateTime,
+            cursorId,
+            newsIds,
+            size,
+        )
     }
 
 }
